requirements.txt
torchvision
numpy
pandas
matplotlib
opencv-python
pytorch
